# swagger/goldfin-service-admin-api.yaml
swagger: '2.0'
info:
  description: REST API for Goldfin Intelligent Invoice Processing
  version: 1.0.0
  title: Goldfin Service API
  termsOfService: 'http://goldfin.io/terms/'
  contact:
    email: info@goldfin.io
host: api.goldfin.io
basePath: /api/v1
...
schemes:
  - https
securityDefinitions:
  SessionKey:
    type: apiKey
    in: header
    name: vnd.io.goldfin.session
  ApiKey:
    type: apiKey
    in: header
    name: vnd.io.goldfin.apikey
security:
  - SessionKey: []
  - ApiKey: []
paths:
...
  /data:
    post:
      tags:
        - inventory
      summary: Upload data series
      description: Upload data series in a file for analysis
      operationId: dataCreate
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: description
          in: formData
          description: An optional description of the data series
          type: string
        - name: tags
          in: formData
          description: Optional tags that apply to this entity passed as a JSON string containing name-value pairs.
          type: string
        - name: process
          in: formData
          description: Optional flag to kick off processing automatically if true
          type: boolean
          default: true
        - name: file
          in: formData
          description: Data series content
          required: true
          type: file
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/DataSeries'
        '400':
          description: Creation failed
          schema:
            $ref: '#/definitions/ApiResponse'
    get:
      tags:
        - inventory
      summary: List data serties
      description: Return a list of metadata entries for all data series
      operationId: dataShowAll
      produces:
        - application/json
      responses:
        '200':
          description: Successful query
          schema:
            type: array
            items:
              $ref: '#/definitions/DataSeries'
  '/data/{id}':
    get:
      tags:
        - inventory
      summary: Return data series metadata
      description: Download data series metadata without content
      operationId: dataShow
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Series ID
          required: true
          type: string
      responses:
        '200':
          description: Successful
          schema:
            $ref: '#/definitions/DataSeries'
        '404':
          description: Not Found
    put:
      tags:
        - inventory
      summary: Update a data series
      description: Update data series description and/or tags.  Other fields are ignored if included in the body.
      operationId: dataUpdate
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: Data series ID
          required: true
          type: string
        - name: body
          in: body
          description: Data series parameters
          schema:
            $ref: '#/definitions/DataSeries'
      responses:
        '200':
          description: Successful
        '404':
          description: Not Found
    delete:
      tags:
        - inventory
      summary: Delete a data series
      description: Delete a data series and any derived information
      operationId: dataDelete
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Series ID
          required: true
          type: string
      responses:
        '200':
          description: Successful
        '404':
          description: Not Found
  '/data/{id}/content':
    get:
      tags:
        - inventory
      summary: Return data series content
      description: Download data series content
      operationId: dataShowContent
      produces:
        - application/octet-stream
      parameters:
        - name: id
          in: path
          description: Series ID
          required: true
          type: string
      responses:
        '200':
          description: Successful
        '404':
          description: Not Found
  '/data/{id}/process':
    post:
      tags:
        - inventory
      summary: Kick off background processing of data series
      description: >-
        Run background processing of data series, which may generate one or more
        inventory records.
      operationId: dataProcess
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Series ID
          required: true
          type: string
      responses:
        '202':
          description: Accepted
        '404':
          description: Not Found
...
definitions:
...
  DataSeries:
    description: A set of one or more observations for analysis
    type: object
    properties:
      id:
        type: string
        description: Series ID
        format: uuid
      name:
        type: string
        description: Name of the source file if known
      description:
        type: string
        description: Optional description of the series
      contentType:
        type: string
        description: 'Internet media type (e.g., application/json)'
      contentLength:
        type: number
        format: int64
        description: Content length in bytes
      thumbprint:
        type: string
        description: SHA-256 thumbprint of content
      locator:
        type: string
        format: url
        description: Storage locator for content
      state:
        description: 'The current processing state of the content.  '
        enum:
          - CREATED
          - PROCESS_REQUESTED
          - PROCESSED
          - ERROR
      format:
        description: 'Kind of data series, e.g., observation'
        enum:
          - OBSERVATION
          - UNKNOWN
      tags:
        $ref: '#/definitions/TagSet'
      creationDate:
        type: string
        description: Date data was uploaded
...
  TagSet:
    type: array
    description: Collection of tags.  If keys are repeated in the array the last value in order takes precedences
    items:
      $ref: '#/definitions/Tag'
  Tag:
    type: object
    description: A key-value pair used to add relevant descriptive information. Tags are not interpreted and may contain data like business unit IDs, contract names, or other information users find relevant to classify or associate the data to which they apply. 
    properties:
      name:
        type: string
        description: A user-defined name for the tag
      value:
        type: string
        description: A user-defined value for the tag
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
...